Redo all the options in AceConfig-3.0
It's worse, but at least I don't have to do it or maintain it.

Had to rename a bunch of the options in the DB, so everyone is starting from
scratch. Lucky this is a beta, because everyone coming direct from 9.x is
starting from scratch anyway

Don't include any libs in the repo, get the packager to fetch and package them.
There's just too many with Ace3 to be sure I wouldn't miss a bugfix.

I quote the words of the great programmers' lament:

  We do these things not because they are easy, but because we thought that
  they would be easy.